Separate out Goodreads search from Author/Book data
