+++ src/NzbDrone.Core/MetadataSource/GoodreadsSearchProxy/Resources/SearchJsonResource.cs
